Web: Refactoring.
