+++ packages/client/src/components/EmptyContainer/RootFolderContainer.js
@@ -2,9 +2,8 @@
 import PersonSvgUrl from "PUBLIC_DIR/images/person.svg?url";
 import PlusSvgUrl from "PUBLIC_DIR/images/plus.svg?url";
 import RoomsReactSvgUrl from "PUBLIC_DIR/images/rooms.react.svg?url";
-import React, { useEffect } from "react";
+
 import { useNavigate, useLocation } from "react-router-dom";
-import styled from "styled-components";
 import { FolderType, RoomSearchArea } from "@docspace/common/constants";
 import { inject, observer } from "mobx-react";
 import { withTranslation, Trans } from "react-i18next";
@@ -13,14 +12,12 @@ import Link from "@docspace/components/link";
 import Text from "@docspace/components/text";
 import Box from "@docspace/components/box";
 import IconButton from "@docspace/components/icon-button";
-import Loaders from "@docspace/common/components/Loaders";
 import RoomsFilter from "@docspace/common/api/rooms/filter";
 import FilesFilter from "@docspace/common/api/files/filter";
 
 import { getCategoryUrl } from "SRC_DIR/helpers/utils";
 import { CategoryType } from "SRC_DIR/helpers/constants";
 
-import PlusIcon from "PUBLIC_DIR/images/plus.react.svg";
 import EmptyScreenPersonalUrl from "PUBLIC_DIR/images/empty_screen_personal.svg?url";
 import EmptyScreenPersonalDarkUrl from "PUBLIC_DIR/images/empty_screen_personal_dark.svg?url";
 import EmptyScreenCorporateSvgUrl from "PUBLIC_DIR/images/empty_screen_corporate.svg?url";
@@ -36,13 +33,7 @@ import EmptyScreenTrashSvgDarkUrl from "PUBLIC_DIR/images/empty_screen_trash_dar
 import EmptyScreenArchiveUrl from "PUBLIC_DIR/images/empty_screen_archive.svg?url";
 import EmptyScreenArchiveDarkUrl from "PUBLIC_DIR/images/empty_screen_archive_dark.svg?url";
 
-import { showLoader, hideLoader } from "./EmptyFolderContainerUtils";
-
-const StyledPlusIcon = styled(PlusIcon)`
-  path {
-    fill: #657077;
-  }
-`;
+import CommonButtons from "./sub-components/CommonButtons";
 
 const RootFolderContainer = (props) => {
   const {
@@ -249,62 +240,7 @@ const RootFolderContainer = (props) => {
   );
 
   const commonButtons = (
-    <span>
-      <div className="empty-folder_container-links">
-        <StyledPlusIcon
-          className="plus-document empty-folder_container-image"
-          data-format="docx"
-          onClick={onCreate}
-          alt="plus_icon"
-        />
-
-        <Box className="flex-wrapper_container">
-          <Link
-            id="document"
-            data-format="docx"
-            onClick={onCreate}
-            {...linkStyles}
-          >
-            {t("Document")},
-          </Link>
-          <Link
-            id="spreadsheet"
-            data-format="xlsx"
-            onClick={onCreate}
-            {...linkStyles}
-          >
-            {t("Spreadsheet")},
-          </Link>
-          <Link
-            id="presentation"
-            data-format="pptx"
-            onClick={onCreate}
-            {...linkStyles}
-          >
-            {t("Presentation")},
-          </Link>
-          <Link
-            id="form-template"
-            data-format="docxf"
-            onClick={onCreate}
-            {...linkStyles}
-          >
-            {t("Translations:NewForm")}
-          </Link>
-        </Box>
-      </div>
-
-      <div className="empty-folder_container-links">
-        <StyledPlusIcon
-          className="plus-folder empty-folder_container-image"
-          onClick={onCreate}
-          alt="plus_icon"
-        />
-        <Link id="folder" {...linkStyles} onClick={onCreate}>
-          {t("Folder")}
-        </Link>
-      </div>
-    </span>
+    <CommonButtons onCreate={onCreate} linkStyles={linkStyles} isRoot />
   );
 
   const trashButtons = (
